refactor: make provision_licenses part of SubscriptionPlan model + resolve circular deps issue
